refactor: move build logic to forc-pkg
