+++ forc-pkg/Cargo.toml
@@ -11,12 +11,14 @@ description = "Building, locking, fetching and updating Sway projects as Forc pa
 [dependencies]
 anyhow = "1"
 forc-util = { version = "0.19.1", path = "../forc-util" }
+fuel-crypto = "0.5"
 fuel-tx = "0.13"
 git2 = { version = "0.14", features = ["vendored-libgit2", "vendored-openssl"] }
 petgraph = { version = "0.6", features = ["serde-1"] }
 semver = { version = "1.0", features = ["serde"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_ignored = "0.1"
+serde_json = "1.0"
 sway-core = { version = "0.19.1", path = "../sway-core" }
 sway-types = { version = "0.19.1", path = "../sway-types" }
 sway-utils = { version = "0.19.1", path = "../sway-utils" }
